refactor teams fetch
