+++ src/services/index.js
@@ -1,9 +1,7 @@
 const makeFetch = async (requestData) => {
   try {
     const res = await fetch(requestData);
-
     const result = await res.json();
-
     return result;
   } catch (err) {
     console.log(err.message);
